Update README.md
Headings:
Headings have been used effectively, with sections devoted to the film "Avengers: Endgame" and the series "The Flash" marked as main (#), sub (# #) and sub subsections (# # #) levels respectively.

Bold and Italic Text:
Important names, quotes, and details were penned as bold letter (*bold*) and italic letter *_italic_* in order to grab attention drawn to her/his write up.

Lists:
Unordered lists (- item) were applied to capture key details (such as release year, genre) and ordered lists (1. item) for emphasis of memorable scenes, themes and similar internal content. This arranges information in a manner that is easy comphensible.

Blockquotes for Quotes:
Considered block quoting (> quote) quoting styles for some moving words in the movie and the series as well.

Links:
Added links to the pages of the movie in the IMDb database for easier access to additional information for the readers.

Images:
Supplemented images of posters using img elements such that a picture is provided for every title.

Table:
A simple table was done comparing "Avengers: Endgame" and "The Flash," which helps to easily look up details and compare them.

Code Block:
A code block was also included to incorporate an expository and concise themed element within "The Flash" portraying simple example code on speed with respect to Barry Allen.

Horizontal Lines:
Horizontal lines (--- ) were also provided in order to help break the content into sections enhancing the cleanliness of the overall presentation.